Ecommerce/requirements.txt
aioredis==2.0.1
aiosmtplib==1.1.6
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
bcrypt==3.2.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.0
cryptography==36.0.2
Deprecated==1.2.13
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.1.3
fakeredis==1.7.1
fastapi==0.75.0
fastapi-mail==1.0.6
greenlet==1.1.2
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.6.0
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
PyMySQL==1.0.2
pyparsing==3.0.7
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
python-slugify==6.1.1
requests==2.27.1
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.32
SQLAlchemy-Utils==0.38.2
starlette==0.17.1
starlette-context==0.3.3
stripe==2.70.0
text-unidecode==1.3
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
wrapt==1.14.0
zipp==3.7.0
